Added map and collision and vision enhancements
Added a way to a map in a text file with an arbitrary size, and a way to unpack it and make it usable. also moved the draw calls to separate draw functions so it is easier to turn them off for performance in the future. also added more points but at the players position for the distance calculation so there is no fisheye happening with the 3D part. and added an fps counter